Time: 4 ms (23.59%) | Memory: 8.4 MB (72.13%) - LeetSync

+++ leetcode/7-reverse-integer/reverse-integer.c
@@ -1,5 +1,5 @@
 int reverse(int x){
-    long int ans = 0;
+    long long ans = 0;
 
     while ( x != 0 ) {
         ans = ans * 10 + ( x % 10 );
